fix(tests): update imports for moved weather module
- chatOrchestrator.test.ts: import from contextIntegrations/weather/index
- openMeteoForecastTool.test.ts: import from contextIntegrations/weather/index
- toolExecutionParity.test.ts: import from contextIntegrations/weather/index
